+++ integrationtests/module-zipfile-server/modules/columnchart/_deprecated_i18n_messages.py
@@ -1,0 +1,11 @@
+# This file should not be executed. Workbench extracts obsolete messages from it.
+# Even far into the future, users may see these messages. Translators must
+# translate them.
+#
+# When removing a message or changing its id, move the old message
+# (i.e. the call to `i18n.trans()`) here.
+i18n.trans(
+    "sameAxesError.message",
+    "You cannot plot Y-axis column {column_name} because it is the X-axis column",
+    {"column_name": "A"},
+)