+++ integrationtests/module-zipfile-server/modules/columnchart/CHANGELOG.md
@@ -1,3 +1,19 @@
+2020-03-18.01
+-------------
+
+* (no user-visible change) Use Workbench's new `dataUrl` pattern
+
+2021-03-01.01
+-------------
+
+* Interaction: on hover, highlight the closest X value and all Y values, and
+  show their values in a tooltip.
+* Change font to Roboto, to match Workbench.
+* HTML: nix obsolete styling of the "ellipsis" button and fix its position.
+* Legend: nix top and right padding by reverting to Vega defaults
+* When title is not supplied, do not show "Column Chart". (It looked English
+  in non-English locales.)
+
 2021-01-20.01
 -------------