+++ integrationtests/module-zipfile-server/modules/pythoncode/CHANGELOG.md
@@ -1,4 +1,4 @@
-2020-03-18.01
+2020-03-19.01
 -------------
 
 * (no user-visible change) Use Workbench's new `dataUrl` pattern